fix hpo
